Add XO linter
