+++ .github/workflows/main.yml
@@ -5,12 +5,20 @@ on: [push, pull_request]
 jobs:
   build:
     runs-on: ubuntu-latest
-
+    strategy:
+      fail-fast: false
+      matrix:
+        node-version:
+          - 18
     steps:
-      - uses: actions/checkout@v1
+      - uses: actions/checkout@v3
+      - uses: actions/setup-node@v3
+        with:
+          node-version: ${{ matrix.node-version }}
       - name: build
         run: |
           npm install
+          npm test
           npm run compile
           npx vsce package
           ls -lt *.vsix
