+++ .gitignore
@@ -1,4 +1,5 @@
 out
 node_modules
+.vscode/settings.json
 .vscode-test/
 *.vsix
