+++ package.json
@@ -123,18 +123,22 @@
     "vscode:prepublish": "npm run compile",
     "compile": "esbuild --bundle --sourcemap=external --minify --external:vscode src/extension.ts --outdir=out --platform=node --format=cjs",
     "watch": "esbuild --watch --bundle --sourcemap=external --external:vscode src/extension.ts --outdir=out --platform=node --format=cjs",
-    "test": "npm run compile && node ./node_modules/vscode/bin/test"
+    "test": "xo"
   },
   "devDependencies": {
     "@types/mocha": "^2.2.48",
-    "@types/node": "^15.6.0",
+    "@types/node": "^18.0.6",
     "@types/vscode": "^1.43.0",
-    "vscode-test": "^1.4.0"
+    "vscode-test": "^1.4.0",
+    "xo": "^0.51.0"
   },
   "dependencies": {
     "esbuild": "^0.12.1",
     "lodash-es": "^4.17.21",
     "lodash.debounce": "^4.0.8",
     "vsce": "^1.88.0"
-  }
+  },
+	"xo": {
+		"prettier": true
+	}
 }
