+++ src/extension.ts
@@ -1,42 +1,47 @@
-'use strict';
 import * as vscode from 'vscode';
-import ZigCompilerProvider from './zigCompilerProvider';
-import { zigBuild } from './zigBuild';
-import { ZigFormatProvider, ZigRangeFormatProvider } from './zigFormat';
+import ZigCompilerProvider from './zig-compiler-provider.js';
+import {zigBuild} from './zig-build.js';
+import {ZigFormatProvider, ZigRangeFormatProvider} from './zig-format.js';
 
-const ZIG_MODE: vscode.DocumentFilter = { language: 'zig', scheme: 'file' };
+// eslint-disable-next-line @typescript-eslint/naming-convention
+const ZIG_MODE: vscode.DocumentFilter = {language: 'zig', scheme: 'file'};
 
+// eslint-disable-next-line import/no-mutable-exports
 export let buildDiagnosticCollection: vscode.DiagnosticCollection;
 export const logChannel = vscode.window.createOutputChannel('zig');
-export const zigFormatStatusBar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left);
+export const zigFormatStatusBar = vscode.window.createStatusBarItem(
+	vscode.StatusBarAlignment.Left,
+);
 
 export function activate(context: vscode.ExtensionContext) {
-    let compiler = new ZigCompilerProvider();
-    compiler.activate(context.subscriptions);
-    vscode.languages.registerCodeActionsProvider('zig', compiler);
+	const compiler = new ZigCompilerProvider();
+	compiler.activate(context.subscriptions);
+	vscode.languages.registerCodeActionsProvider('zig', compiler);
 
-    context.subscriptions.push(logChannel);
-    context.subscriptions.push(
-        vscode.languages.registerDocumentFormattingEditProvider(
-            ZIG_MODE,
-            new ZigFormatProvider(logChannel),
-        ),
-    );
+	context.subscriptions.push(
+		logChannel,
+		vscode.languages.registerDocumentFormattingEditProvider(
+			ZIG_MODE,
+			new ZigFormatProvider(logChannel),
+		),
+		vscode.languages.registerDocumentRangeFormattingEditProvider(
+			ZIG_MODE,
+			new ZigRangeFormatProvider(logChannel),
+		),
+	);
 
-    context.subscriptions.push(
-        vscode.languages.registerDocumentRangeFormattingEditProvider(
-            ZIG_MODE,
-            new ZigRangeFormatProvider(logChannel),
-        ),
-    );
-
-    buildDiagnosticCollection = vscode.languages.createDiagnosticCollection('zig');
-    context.subscriptions.push(buildDiagnosticCollection);
-
-    // Commands
-    context.subscriptions.push(vscode.commands.registerCommand('zig.build.workspace', () => zigBuild()));
-    context.subscriptions.push(vscode.commands.registerCommand('zig.format.file', () => console.log('test')));
+	buildDiagnosticCollection =
+		vscode.languages.createDiagnosticCollection('zig');
+	context.subscriptions.push(
+		buildDiagnosticCollection,
+		vscode.commands.registerCommand('zig.build.workspace', () => {
+			zigBuild();
+		}),
+		vscode.commands.registerCommand('zig.format.file', () => {
+			console.log('test');
+		}),
+	);
 }
 
-export function deactivate() {
-}
+// eslint-disable-next-line @typescript-eslint/no-empty-function
+export function deactivate() {}
